+++ CHANGELOG.md
@@ -29,6 +29,7 @@
         - Many changes to the functions, associated methods, and `BlockBuilder` were made to support this new structure.
 
     - Breaking: `apps::Session::new()` (formerly `AllIsCubesAppState::new()`) is now an async function.
+    - Breaking: `apps::StandardCameras` now works with a `ListenableSource<Viewport>` instead of a `Viewport` and `set_viewport()` method.
     - Breaking: `linking::BlockProvider::new()` is now an async function.
     - Breaking: The `linking::BlockModule` trait now requires the [`exhaust::Exhaust`](https://docs.rs/exhaust/latest/exhaust/trait.Exhaust.html) trait in place of `strum::IntoEnumIterator`. This allows implementors to use enums with fields (or non-enums).